cluster-ui: fix sort on tables
Sort on tables of Statement, Transaction and Sessions page were
not using the right value from the url regarding sorting. It was
using the right sortKey, but always doing ascending. Fix done so
it uses the right value for the boolean.

Since the sortKet might be different for different users on Statement
Page (difference on columns being displayed), use the name of the
columns, instead of the key.

Fixes: https://github.com/cockroachdb/cockroach/issues/65085

Release note (bug fix): Sort on tables are now picking up the
correct value from the url.